+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+### Changed
+- Update Interactable System package to 2.0.0
+
 ## [0.3.0] - 2026-03-09
 ### Changed
 - Rename event OnDialogueStopped -> OnDialogueFinished