+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-
+## [0.4.0] - 2026-03-10
 ### Changed
 - Update Interactable System package to 2.0.0
 
@@ -51,7 +51,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/1mbitshorde/DialogueSystem/compare/0.3.0...main
+[Unreleased]: https://github.com/1mbitshorde/DialogueSystem/compare/0.4.0...main
+[0.4.0]: https://github.com/1mbitshorde/DialogueSystem/tree/0.4.0/
 [0.3.0]: https://github.com/1mbitshorde/DialogueSystem/tree/0.3.0/
 [0.2.0]: https://github.com/1mbitshorde/DialogueSystem/tree/0.2.0/
 [0.1.0]: https://github.com/1mbitshorde/DialogueSystem/tree/0.1.0/